pyproject.toml
[project]
name = "karoo-garmin-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "garminconnect>=0.2.19",
    "pyjwt>=2.9.0",
    "requests>=2.32.3",
]
